Make highlightlength a bit more foolproof
HighlightLength will now adapt to the largest element in the menu. As a
result, I've moved the Display method out of the base class as it now
needs to be called after HighlightLength is calculated in HighlightMenu.

Also in the news: Console.CursorTop started working again for some
reason, I further simplified the process of printing highlighted
options, and I added a word into a comment.