Another CursorPos squash & clarification
I only just found out that HighlightMenu wasn't showing numbers even
though I'd enabled it, and despite the fact that this enable was there
for weeks.

So I tried to add numbers, but due to the redrawing code it wasn't
working out as quickly as I'd have liked, so it's scrapped for the
forseeable future. I'm sure people can count, anyway.